Crate axfeat

Top-level feature selection for ArceOS.

Cargo Features

CPU

smp: Enable SMP (symmetric multiprocessing) support.

fp-simd: Enable floating point and SIMD support.

Interrupts:

irq: Enable interrupt handling support.

Memory

alloc: Enable dynamic memory allocation.

alloc-tlsf: Use the TLSF allocator.

alloc-slab: Use the slab allocator.

alloc-buddy: Use the buddy system allocator.

paging: Enable page table manipulation.

tls: Enable thread-local storage.

Task management

multitask: Enable multi-threading support.

sched-fifo: Use the FIFO cooperative scheduler.

sched-rr: Use the Round-robin preemptive scheduler.

sched-cfs: Use the Completely Fair Scheduler (CFS) preemptive scheduler.

Upperlayer stacks (fs, net, display)

fs: Enable file system support.

myfs: Allow users to define their custom filesystems to override the default.

net: Enable networking support.

display: Enable graphics support.

Device drivers

bus-mmio: Use device tree to probe all MMIO devices.

bus-pci: Use PCI bus to probe all PCI devices.

driver-ramdisk: Use the RAM disk to emulate the block device.

driver-ixgbe: Enable the Intel 82599 10Gbit NIC driver.

driver-bcm2835-sdhci: Enable the BCM2835 SDHCI driver (Raspberry Pi SD card).

Logging

log-level-off: Disable all logging.

log-level-error, log-level-warn, log-level-info, log-level-debug, log-level-trace: Keep logging only at the specified level or higher.